# backend/docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Club Booking SaaS Platform API
  description: >
    Multi-tenant booking engine for Players and Club Owners.
    
    ## Authentication
    All protected endpoints use JWT stored in an **HttpOnly cookie** named `access_token_cookie`.
    The browser automatically sends the cookie with each request – no manual token management is required.
    Endpoints that require specific roles (e.g., `owner`, `player`) are marked accordingly.

    ## Implementation Deviations Logged
    1. Operating hours fields (`open_time`, `close_time`, `slot_duration_minutes`) are flat fields on `Club` and `Court` models, not nested objects.
    2. Overlap validation for Bookings is performed at the application layer via Python queries instead of PostgreSQL GiST constraints, as SQLite is used.
    3. Additional modules implemented and documented: Memberships, Events, Payments, Attendance, and Notifications.
  version: 2.1.0
servers:
  - url: http://127.0.0.1:5000/api/v1
    description: Local Flask Development Server
security:
  - cookieAuth: []   # JWT stored in HttpOnly cookie
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT authentication cookie (HttpOnly)
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
paths:
...
  /payments/stripe/create-payment-intent:
    post:
      summary: Create a Stripe PaymentIntent (JWT required)
      description: >
        Records a pending Payment, creates a Stripe PaymentIntent, and returns
        the client_secret for the frontend (Stripe.js) to confirm the card
        payment. The **amount is resolved server-side** from the referenced
        entity (never trusted from the client):
        booking -> flat BOOKING_FEE (booking must belong to the caller and be
        active); membership -> MembershipPlan.price_monthly; event ->
        Event.registration_fee (must be > 0).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_type, reference_id]
              properties:
                payment_type:
                  type: string
                  enum: [booking, membership, event]
                reference_id:
                  type: integer
                  example: 1
                  description: "booking_id, membership plan_id, or event_id depending on payment_type"
                currency: { type: string, example: inr, description: "ISO 4217; defaults to STRIPE_DEFAULT_CURRENCY" }
      responses:
        '201':
          description: PaymentIntent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_id: { type: integer }
                  client_secret: { type: string }
                  amount: { type: number, format: float, description: "resolved server-side" }
                  currency: { type: string }
                  payment_type: { type: string }
                  reference_id: { type: integer }
                  status: { type: string, example: pending }
                  publishable_key: { type: string, nullable: true }
        '400':
          description: Validation error (bad payment_type, or free event)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Not authenticated
        '403':
          description: Booking belongs to another user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Referenced booking / plan / event not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: This item has already been paid for (ALREADY_PAID)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '502':
          description: Stripe gateway error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Stripe not configured (STRIPE_SECRET_KEY missing)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /payments/stripe/webhook:
    post:
      summary: Stripe webhook (public; verified by signature)
      description: >
        Receives Stripe events. The raw request body is verified against the
        `Stripe-Signature` header using STRIPE_WEBHOOK_SECRET.


        On `payment_intent.succeeded` the Payment is marked `completed` **and
        the purchase is fulfilled**:
        `membership` activates the Membership for the paid plan;
        `event` confirms the user's EventRegistration (capacity still enforced);
        `booking` confirms the booking. A Notification is raised in each case.
        On `payment_intent.payment_failed` / `.canceled` the Payment is marked
        `failed` and nothing is fulfilled.


        Fulfilment is **idempotent** — Stripe retries are safe; a repeat event
        returns `fulfilment: already_completed`. A late failure event can never
        downgrade an already-completed payment.
      security: []
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe event payload
      responses:
        '200':
          description: Event processed (also returned for unknown payments so Stripe stops retrying)
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean }
                  type: { type: string, example: payment_intent.succeeded }
                  payment_id: { type: integer }
                  fulfilment:
                    type: string
                    description: Outcome of fulfilment
                    enum: [membership_activated, membership_already_active,
                           registration_confirmed, registration_already_confirmed,
                           event_full, booking_confirmed, booking_inactive,
                           already_completed, already_failed, failed,
                           no_matching_payment, no_handler, ignored]
        '400':
          description: Invalid payload or signature
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: Stripe webhook not configured
  /payments/stripe/status/{paymentId}:
    get:
      summary: Poll a payment's status and fulfilment result (JWT required)
      description: >
        Used by the client after confirming the card with Stripe.js. Once the
        webhook has been processed the status becomes `completed`, and for
        memberships/events the resulting `membership_id` / `registration_status`
        is included so the client can route the user onward. A user may only
        read their own payments.
      parameters:
        - name: paymentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_id: { type: integer }
                  status: { type: string, enum: [pending, completed, failed, refunded] }
                  amount: { type: number, format: float }
                  currency: { type: string }
                  payment_type: { type: string, enum: [booking, membership, event] }
                  reference_id: { type: integer }
                  transaction_id: { type: string, nullable: true }
                  created_at: { type: string, format: date-time }
                  membership_id:
                    type: integer
                    nullable: true
                    description: Present for completed membership payments
                  registration_status:
                    type: string
                    nullable: true
                    description: Present for completed event payments
        '403':
          description: Not authorized to view this payment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Payment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ================= ATTENDANCE =================
...